frontend

+++ backend/controllers/comment.js
@@ -34,16 +34,14 @@ exports.deleteComment = (req, res, next ) => {
 
   Comment.findOne({ where: { id }})
   .then((comment) => {
+    console.log(comment)
+    if (!comment) {
+      return res.status(404).json({ message: "pas de commentaire!"})
+    }
     if (comment.UserId === userId || isAdmin) {
-      /*Post.findOne({ where: {id: comment.postId}})
-        .then((post) => {
-          post.update({ where: {id: comment.postId}})
-            .then(() => res.status(200).json({ message: "Commentaire supprimé ! "}))
-            .catch(error => res.status(400).json({ error }))
-        });*/
-          comment.destroy({ where: { id }})
-            .then(() => res.status(200).json({ message: "commentaire supprimé ! "}))
-            .catch(error => res.status(400).json({ error }));
+      comment.destroy({ where: { id }})
+        .then(() => res.status(200).json({ message: "commentaire supprimé ! "}))
+        .catch(error => res.status(400).json({ error }));
     } else {
       return res.status(401).json({error});
     }
@@ -51,11 +49,17 @@ exports.deleteComment = (req, res, next ) => {
   .catch(error => res.status(500).json({ error }));
 };
 
-exports.getAllComments = (req, res, next ) => {
+exports.getOneComment = (req, res, next ) => {
   Comment.findAll({ where: {postId: req.params.id }, 
     include: [ {model: User} ],
     order: [["createdAt", "DESC"]]
   })
+    .then(comment => res.status(200).json(comment))
+    .catch(error => res.status(400).json({error}));
+}
+
+exports.getAllComments = (req, res, next ) => {
+  Comment.findAll()
     .then(comment => res.status(200).json(comment))
     .catch(error => res.status(400).json({error}));
 }
